Move Router and Handler into combined package
