+++ src/main/java/com/heatheryou/httpserver/constants/EntityHeader.java
@@ -4,4 +4,6 @@ public class EntityHeader {
     public static final String ALLOWED_METHODS = "Allow:";
     public static final String CONTENT_LENGTH = "Content-Length:";
     public static final int CONTENT_LENGTH_0 = 0;
+    public static final String LOCATION = "Location:";
+    public static final String SIMPLE_GET_URL = "http://0.0.0.0:5000/simple_get";
 }
