+++ src/main/java/com/heatheryou/httpserver/route/RequestHandler.java
@@ -1,4 +1,4 @@
-package com.heatheryou.httpserver.handler;
+package com.heatheryou.httpserver.route;
 
 import com.heatheryou.httpserver.Request;
 import com.heatheryou.httpserver.Response;
