+++ src/test/java/com/heatheryou/httpserver/route/handler/NoRouteFoundHandlerTest.java
@@ -1,4 +1,4 @@
-package com.heatheryou.httpserver.handler;
+package com.heatheryou.httpserver.route.handler;
 
 import com.heatheryou.httpserver.Request;
 import com.heatheryou.httpserver.Response;
